stream/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
mutagen==1.45.1
pycryptodomex==3.11.0
requests==2.26.0
urllib3==1.26.7
yt-dlp==2023.06.22
python-dotenv==0.19.2
